+++ compiler-core/src/type_/expression.rs
@@ -2097,7 +2097,9 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
                 .ok_or_else(|| Error::UnknownModule {
                     name: module_alias.clone(),
                     location: *module_location,
-                    imported_modules: self.environment.imported_modules.keys().cloned().collect(),
+                    suggestions: self
+                        .environment
+                        .suggest_modules(module_alias, Imported::Value(label.clone())),
                 })?;
 
             let constructor =
@@ -2379,7 +2381,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
 
     fn infer_value_constructor(
         &mut self,
-        module: &Option<EcoString>,
+        module: &Option<(EcoString, SrcSpan)>,
         name: &EcoString,
         location: &SrcSpan,
     ) -> Result<ValueConstructor, Error> {
@@ -2399,20 +2401,17 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
             }
 
             // Look in an imported module for a binding with this name
-            Some(module_name) => {
+            Some((module_name, module_location)) => {
                 let (_, module) = &self
                     .environment
                     .imported_modules
                     .get(module_name)
                     .ok_or_else(|| Error::UnknownModule {
-                        location: *location,
+                        location: *module_location,
                         name: module_name.clone(),
-                        imported_modules: self
+                        suggestions: self
                             .environment
-                            .imported_modules
-                            .keys()
-                            .cloned()
-                            .collect(),
+                            .suggest_modules(module_name, Imported::Value(name.clone())),
                     })?;
                 module
                     .values
@@ -2516,7 +2515,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
                 ..
             } if args.is_empty() => {
                 // Register the module as having been used if it was imported
-                if let Some(ref module) = &module {
+                if let Some((module, _)) = &module {
                     _ = self.environment.unused_modules.remove(module);
                     _ = self.environment.unused_module_aliases.remove(module);
                 }
@@ -2561,7 +2560,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
                 ..
             } => {
                 // Register the module as having been used if it was imported
-                if let Some(ref module) = &module {
+                if let Some((module, _)) = &module {
                     _ = self.environment.unused_modules.remove(module);
                     _ = self.environment.unused_module_aliases.remove(module);
                 }
@@ -2591,7 +2590,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
                 // TODO: resvisit this. It is rather awkward at present how we
                 // have to convert to this other data structure.
                 let fun = match &module {
-                    Some(module_alias) => {
+                    Some((module_alias, _)) => {
                         let type_ = Arc::clone(&constructor.type_);
                         let module_name = self
                             .environment
@@ -2688,7 +2687,7 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
                 ..
             } => {
                 // Register the module as having been used if it was imported
-                if let Some(ref module) = &module {
+                if let Some((module, _)) = &module {
                     _ = self.environment.unused_modules.remove(module);
                     _ = self.environment.unused_module_aliases.remove(module);
                 }